Fix figure error (#56)
* Fix figure error

* generator error

+++ src/generator/headers/Figure.h
@@ -17,11 +17,11 @@ class Figure {
         Figure(vector<Point*>);
         long getNPoints(){ return npoints; };
         vector<Point*> getPoints(){ return points; };
-        void createPlane(double size);
-        void createBox(double x, double y, double z, int d);
-        void createSphere(double radius, int slices, int stacks);
-        void createCone(double base, double height, int slices, int stacks);
-        void createCylinder(double r, double height, int slices, int stacks);
+        void createPlane(float size);
+        void createBox(float x, float y, float z, int d);
+        void createSphere(float radius, int slices, int stacks);
+        void createCone(float base, float height, int slices, int stacks);
+        void createCylinder(float r, float height, int slices, int stacks);
 };
 
 #endif
